spec: pull virt-v2v for Vdsm
To enable the external VM import feature in Vdsm, users needs
to install the virt-v2v package.
This package  is optional, meaning that all other features will
work as usual if virt-v2v is not present, still we
want this enabled by default on RHEL/RHEV.

To make things easier and let this flow work out of the box,
we pull virt-v2v as explicit dependency.

This is not needed by safelease, and it is another hack to
add on the existing one. It will go away as soon as Vdsm
can pull its arch-dependent dependencies.

Change-Id: Iec2596e11c829bd3b4ef47585a1836e6dd282870
Bug-Url: https://bugzilla.redhat.com/1318014
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54857
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Continuous-Integration: Yaniv Bronhaim <ybronhei@redhat.com>